Update SharedFunctions and Tetsouo_BLM
In SharedFunctions.lua, the condition for checking the TreasureMode state has been simplified. The new condition is:
if state.TreasureMode.value == 'SATA' or state.TreasureMode.value == 'Fulltime' then
This change removes the redundant check for the existence of state.TreasureMode.

In Tetsouo_BLM.lua, several functions that were previously commented out for testing purposes have been uncommented and are now active again.
This restores the full functionality of the Black Mage job script.