Correct the name of 'Assassin's Culottes' in THF_S
ET.lua

In this commit, we have made a minor but important correction to the
`THF/THF_SET.lua` file. The name of 'Assassin's Culottes' was previously
written in lowercase as 'assassin's culottes'. This has been corrected
to 'Assassin's Culottes' to match the exact item name in the game. This
ensures that the item can be correctly identified and used within the
Steal set.